+++ packages/brick_offline_first/lib/src/offline_queue/request_sqlite_cache.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'package:path/path.dart' as p;
 import 'package:http/http.dart' as http;
 import 'package:logging/logging.dart';
 import 'package:sqflite/sqflite.dart';
@@ -8,7 +7,6 @@ import 'package:brick_offline_first/src/offline_queue/request_sqlite_cache_manag
 /// Serialize and Deserialize a [http.Request] from SQLite.
 class RequestSqliteCache {
   final http.Request request;
-  final String databaseName;
 
   /// Matches any HTTP requests that send data (or 'push'). 'Pull' requests most often have an
   /// outcome that exists in memory (e.g. deserializing to a model). Since callbacks cannot
@@ -16,10 +14,7 @@ class RequestSqliteCache {
   /// disposal or a crash has since occurred), 'pull' requests will be ignored.
   bool get requestIsPush => ['POST', 'PUT', 'DELETE', 'PATCH'].contains(request.method);
 
-  RequestSqliteCache(
-    this.request,
-    this.databaseName,
-  );
+  RequestSqliteCache(this.request);
 
   /// Builds request into a new SQLite-insertable row
   /// Only available if [request] was initialized from [fromRequest]
@@ -39,9 +34,8 @@ class RequestSqliteCache {
   }
 
   /// Removes the request from the database and thus the queue
-  Future<int> delete() async {
-    final db = await _getDb();
-    final response = await _findRequestInDatabase();
+  Future<int> delete(Database db) async {
+    final response = await _findRequestInDatabase(db);
 
     if (response != null && response.isNotEmpty) {
       return await db.transaction((txn) async {
@@ -58,15 +52,14 @@ class RequestSqliteCache {
 
   /// If the request already exists in the database, increment attemps and
   /// set `updated_at` to current time.
-  Future<int> insertOrUpdate(Logger logger) async {
-    final db = await _getDb();
-    final response = await _findRequestInDatabase();
+  Future<int> insertOrUpdate(Database db, {Logger logger}) async {
+    final response = await _findRequestInDatabase(db);
 
     return db.transaction((txn) async {
       if (response == null || response.isEmpty) {
         final serialized = toSqlite();
 
-        logger.fine('adding to queue: $serialized');
+        logger?.fine('adding to queue: $serialized');
         return await txn.insert(
           HTTP_JOBS_TABLE_NAME,
           serialized,
@@ -74,7 +67,7 @@ class RequestSqliteCache {
       }
       final methodWithUrl =
           [response[HTTP_JOBS_REQUEST_METHOD_COLUMN], response[HTTP_JOBS_URL_COLUMN]].join(' ');
-      logger.warning(
+      logger?.warning(
           'failed, attempt #${response[HTTP_JOBS_ATTEMPTS_COLUMN]} in $methodWithUrl : $response');
       return await txn.update(
         HTTP_JOBS_TABLE_NAME,
@@ -89,9 +82,7 @@ class RequestSqliteCache {
     });
   }
 
-  Future<Map<String, dynamic>> _findRequestInDatabase() async {
-    final db = await _getDb();
-
+  Future<Map<String, dynamic>> _findRequestInDatabase(Database db) async {
     final columns = [
       HTTP_JOBS_BODY_COLUMN,
       HTTP_JOBS_ENCODING_COLUMN,
@@ -110,14 +101,6 @@ class RequestSqliteCache {
     );
 
     return response?.isNotEmpty == true ? response.first : null;
-  }
-
-  Database _db;
-  Future<Database> _getDb() async {
-    if (_db?.isOpen == true) return _db;
-    final databasesPath = await getDatabasesPath();
-    final path = p.join(databasesPath, databaseName);
-    return _db = await openDatabase(path);
   }
 
   /// Recreate a request from SQLite data