+++ packages/brick_sqlite/lib/testing.dart
@@ -1,1 +1,0 @@
-export 'package:brick_sqlite/src/testing/stub_sqlite.dart';