+++ packages/brick_sqlite/test/__mocks__.dart
@@ -1,7 +1,7 @@
-import 'package:brick_sqlite_abstract/db.dart' show InsertTable;
+import 'package:brick_sqlite_abstract/db.dart';
 import 'package:brick_sqlite/sqlite.dart';
 
-const sqliteTableName = "DemoModel";
+const sqliteTableName = 'DemoModel';
 
 class DemoModelAssoc extends SqliteModel {
   DemoModelAssoc(this.name);
@@ -18,7 +18,8 @@ class DemoModelAdapter with SqliteAdapter<DemoModel> {
   final tableName = sqliteTableName;
 
   Future<DemoModel> fromSqlite(map, {provider, repository}) {
-    final composedModel = DemoModel(map['name'])..primaryKey = map[InsertTable.PRIMARY_KEY_COLUMN];
+    final composedModel = DemoModel(map['full_name'])
+      ..primaryKey = map[InsertTable.PRIMARY_KEY_COLUMN];
     return Future.value(composedModel);
   }
 
@@ -49,15 +50,37 @@ class DemoModelAdapter with SqliteAdapter<DemoModel> {
   };
 
   Future<Map<String, dynamic>> toSqlite(instance, {provider, repository}) {
-    return Future.value({'name': instance.name});
+    return Future.value({'full_name': instance.name});
   }
 
   primaryKeyByUniqueColumns(instance, db, {provider, repository}) => null;
 }
 
+const _demoModelMigrationCommands = [
+  InsertTable('DemoModel'),
+  InsertTable('DemoModelAssoc'),
+  InsertColumn('id', Column.integer, onTable: 'DemoModel', unique: true),
+  InsertColumn('simple_bool', Column.boolean, onTable: 'DemoModel'),
+  InsertColumn('last_name', Column.varchar, onTable: 'DemoModel'),
+  InsertColumn('full_name', Column.varchar, onTable: 'DemoModel'),
+  InsertColumn('many_assoc', Column.varchar, onTable: 'DemoModel'),
+  InsertForeignKey('DemoModel', 'DemoModelAssoc',
+      foreignKeyColumn: 'assoc_DemoModelAssoc_brick_id'),
+  InsertColumn('complex_field_name', Column.varchar, onTable: 'DemoModel'),
+];
+
+class DemoModelMigration extends Migration {
+  const DemoModelMigration()
+      : super(
+          version: 2,
+          up: _demoModelMigrationCommands,
+          down: _demoModelMigrationCommands,
+        );
+}
+
 class DemoModelAssocAdapter extends DemoModelAdapter {
   DemoModelAssocAdapter();
-  final tableName = sqliteTableName + "Assoc";
+  final tableName = sqliteTableName + 'Assoc';
 }
 
 final Map<Type, SqliteAdapter<SqliteModel>> _mappings = {